Revert "Preliminary refactor for supporting spans over HTTP (#176)"
This reverts commit 51df6ab195fb5c7ac83610fe8c33cade19172d4c.
