+++ crossdock/server/endtoend.py
@@ -27,7 +27,6 @@ from jaeger_client.constants import (
     SAMPLER_TYPE_REMOTE,
 )
 from jaeger_client.sampler import RemoteControlledSampler, ConstSampler
-from jaeger_client.senders import UDPSender
 from jaeger_client.reporter import Reporter
 from jaeger_client.tracer import Tracer
 
@@ -66,13 +65,8 @@ class EndToEndHandler(object):
         init_sampler = cfg.sampler
         channel = self.local_agent_sender
 
-        sender = UDPSender(
-            io_loop=channel.io_loop,
-            host=os.getenv('AGENT_HOST', 'jaeger-agent'),
-            port=cfg.local_agent_reporting_port,
-        )
         reporter = Reporter(
-            sender=sender,
+            channel=channel,
             flush_interval=cfg.reporter_flush_interval)
 
         remote_sampler = RemoteControlledSampler(
